+++ src/main/java/br/edu/ifnmg/coordenador/Coordenador.java
@@ -14,8 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package br.edu.ifnmg.sigec.entity;
+package br.edu.ifnmg.coordenador;
 
+import br.edu.ifnmg.curso.Curso;
 import java.io.Serializable;
 import java.util.ArrayList;
 import javax.persistence.Column;
