+++ src/main/resources/META-INF/persistence.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <persistence version="2.2" xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_2.xsd">
   <!-- Define Persistence Unit -->
-  <persistence-unit name="SistemaPU">
+  <persistence-unit name="SistemaPU" transaction-type="JTA">
     <jta-data-source>java:/SistemaDS</jta-data-source>
-    <class>br.edu.ifnmg.sigec.entity.Certificado</class>
-    <class>br.edu.ifnmg.sigec.entity.Coordenador</class>
-    <class>br.edu.ifnmg.sigec.entity.Curso</class>
-    <class>br.edu.ifnmg.sigec.entity.Estudante</class>
+    <class>br.edu.ifnmg.estudante.Estudante</class>
+    <class>br.edu.ifnmg.certificado.Certificado</class>
+    <class>br.edu.ifnmg.curso.Curso</class>
+    <class>br.edu.ifnmg.coordenador.Coordenador</class>
     <properties>
       <property name="javax.persistence.schema-generation.database.action" value="drop-and-create"/>
     </properties>
